pyproject.toml
[tool.poetry]
name = "myassistant"
version = "0.1.0"
description = "My Personal Assistant - custom AI based on Chainlit, llama-index and Antrhopic Claude 3."
authors = ["Mathijs de Bruin <mathijs@mathijsfietst.nl>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
chainlit = "^1.0.500"
llama-index-readers-file = "^0.1.13"
llama-index-embeddings-fastembed = "^0.1.4"
llama-index-llms-anthropic = "^0.1.7"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^8.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
